+++ src/Entity/Flickity.php
@@ -7,6 +7,7 @@
 namespace Drupal\flickity\Entity;
 
 
+use Drupal\Component\Serialization\Yaml;
 use Drupal\Core\Config\Entity\ConfigEntityBase;
 
 /**
@@ -43,19 +44,72 @@ class Flickity extends ConfigEntityBase implements FlickityInterface {
 
   /**
    * The machine name of this Flickity configuration.
+   *
    * @var string
    */
   protected $id;
 
   /**
    * The human-readable name of this Flickity configuration.
+   *
    * @var string
    */
   protected $label;
 
   /**
    * An array of Flickity options.
+   *
    * @var array
    */
   protected $options;
+
+
+  /**
+   * Gets the options.
+   *
+   * @return array
+   *   The options array.
+   */
+  public function getOptions() {
+
+    return $this->options;
+  }
+
+  /**
+   * Gets the options formatted as Flickity options.
+   *
+   * @return array
+   *   The options array ready to use as Flickity options.
+   */
+  public function getFormattedOptions() {
+
+    $options = $this->options;
+
+    if(!empty($options['advanced'])) {
+      array_merge($options, Yaml::decode($options['advanced']));
+      unset($options['advanced']);
+    }
+
+    return $options;
+  }
+
+  /**
+   * Gets the configuration list.
+   *
+   * @return array
+   *   An array of Flickity configuration. The config ID is the key, and the
+   *   config label the value.
+   */
+  public static function getConfigList() {
+
+    $entities = Flickity::loadMultiple();
+    $config_list = array();
+
+    /** @var FlickityInterface $entity */
+    foreach ($entities as $entity) {
+      $config_list[$entity->get('id')] = $entity->get('label');
+    }
+
+    return $config_list;
+  }
 }