+++ CHANGELOG.md
@@ -5,3 +5,5 @@ All notable changes to this project will be documented in this file.
 ## [v0.1.0]
 
 - Downloaded Flickity v2.0.5, added libraries.yml file. Flickity config entity.
+
+- Advanced options in Yaml format. Flickity field formatter.